create enums folder and move skill type enum to the folder
