+++ skill-tree/src/main/java/com/RDS/skilltree/enums/SkillTypeEnum.java
@@ -1,4 +1,4 @@
-package com.RDS.skilltree.Skill;
+package com.RDS.skilltree.enums;
 
 public enum SkillTypeEnum {
     ATOMIC,
